Organize link updating
